--- applicant/templates/applicant/login.html
@@ -1,24 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <title>My Site</title>
-</head>
-<body>
-    <header>
-        <h1>My Site</h1>
-    </header>
-    <nav>
-        <a href="{% url 'register' %}">Register</a>
-        <a href="{% url 'login' %}">Login</a> 
-        <a href="{% url 'logout' %}">Logout</a>
-    </nav>
-    <main>
-      <h2>Login</h2>
-      <form method="post">
-        {% csrf_token %}
-        {{ form.as_p }}
-        <button type="submit">Login</button>
-      </form>
-    </main>
-</body>
-</html>
