--- applicant/templates/applicant/logout.html
@@ -1,20 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <title>My Site</title>
-</head>
-<body>
-    <header>
-        <h1>My Site</h1>
-    </header>
-    <nav>
-        <a href="{% url 'register' %}">Register</a>
-        <a href="{% url 'login' %}">Login</a> 
-        <a href="{% url 'logout' %}">Logout</a>
-    </nav>
-    <main>
-      <h2>Logout</h2>
-      <p>You have been logged out. <a href="{% url 'login' %}">Login again</a></p>
-    </main>
-</body>
-</html>
