--- applicant/templates/applicant/profile.html
@@ -1,24 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <title>My Site</title>
-</head>
-<body>
-    <header>
-        <h1>My Site</h1>
-    </header>
-    <nav>
-        <a href="{% url 'register' %}">Register</a>
-        <a href="{% url 'login' %}">Login</a> 
-        <a href="{% url 'logout' %}">Logout</a>
-    </nav>
-    <main>
-      <h2>Profile</h2>
-      <form method="POST">
-        {% csrf_token %}
-        {{ p_form.as_p }}
-        <button type="submit">Update Profile</button>
-      </form>
-    </main>
-</body>
-</html>
