--- applicant/templates/applicant/register.html
@@ -1,25 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <title>My Site</title>
-</head>
-<body>
-    <header>
-        <h1>My Site</h1>
-    </header>
-    <nav>
-        <a href="{% url 'register' %}">Register</a>
-        <a href="{% url 'login' %}">Login</a> 
-        <a href="{% url 'logout' %}">Logout</a>
-    </nav>
-    <main>
-      <h2>Register</h2>
-      <form method="POST">
-        {% csrf_token %}
-        {{ form.as_p }}
-        <button type="submit">Register</button>
-      </form>
-      <p>Already have an account? <a href="{% url 'login' %}">Login</a></p>
-    </main>
-</body>
-</html>
