+++ placement_cell_be/settings.py
@@ -12,7 +12,7 @@ https://docs.djangoproject.com/en/5.0/ref/settings/
 import os
 from pathlib import Path
 
-from dotenv import load_dotenv 
+from dotenv import load_dotenv
 load_dotenv()
 
 # Build paths inside the project like this: BASE_DIR / 'subdir'.
@@ -40,9 +40,7 @@ INSTALLED_APPS = [
     "django.contrib.sessions",
     "django.contrib.messages",
     "django.contrib.staticfiles",
-    "accounts.apps.AccountsConfig",
     "rest_framework",
-    "applicant",
 ]
 
 EXTERNAL_APPS = [
@@ -122,8 +120,6 @@ DATABASES = {
         "PORT": '5432',
     }
 }
-
-
 
 
 # Password validation
